same thing as before but this time define WebSocketMessageReader: * also rexport WebSocketMessageReader sp client lib can create it. * move message-writers to new folder.
builds fine.
