+++ jsonrpc/src/main.ts
@@ -17,8 +17,8 @@ import { Message, MessageType,
 	NotificationType5, NotificationType6, NotificationType7, NotificationType8, NotificationType9
 } from './messages';
 
-import { MessageReader, DataCallback, StreamMessageReader, IPCMessageReader } from './messageReader';
-import { MessageWriter, StreamMessageWriter, IPCMessageWriter } from './messageWriter';
+import { MessageReader, DataCallback, StreamMessageReader, IPCMessageReader, WebSocketMessageReader } from './messageReader';
+import { MessageWriter, StreamMessageWriter, IPCMessageWriter, WebSocketMessageWriter } from './messageWriter';
 import { Disposable, Event, Emitter } from './events';
 import { CancellationTokenSource, CancellationToken } from './cancellation';
