+++ jsonrpc/src/message-writers/message-writer.ts
@@ -1,0 +1,8 @@
+import { Event, Emitter } from '../events';
+import { Message } from '../messages';
+
+export interface MessageWriter {
+	onError: Event<[Error, Message, number]>;
+	onClose: Event<void>;
+	write(msg: Message): void;
+}
